Comparison error fix

+++ ishmuratov.eric/P4/numrow_check.cpp
@@ -4,10 +4,10 @@ unsigned int ishmuratov::checkRow(const int * matrix, size_t rows, size_t cols)
 {
   size_t max_seq = 0;
   unsigned int row_num = 0;
-  for (int i = 0; i < rows; ++i)
+  for (size_t i = 0; i < rows; ++i)
   {
     size_t cur_seq = 1;
-    for (int j = 0; j < cols - 1; ++j)
+    for (size_t j = 0; j < cols - 1; ++j)
     {
       if (matrix[i * cols + j] == matrix[i * cols + j + 1])
       {
