+++ ishmuratov.eric/P4/trimatrix_check.cpp
@@ -7,9 +7,9 @@ bool ishmuratov::lowerMatrixCheck(const int * matrix, size_t rows, size_t cols)
   {
     isLower = false;
   }
-  for (int i = 0; i < rows; ++i)
+  for (size_t i = 0; i < rows; ++i)
   {
-    for (int j = i + 1; j < cols; ++j)
+    for (size_t j = i + 1; j < cols; ++j)
     {
       if (matrix[i * cols + j] != 0)
       {
